handle SummerSession
