+++ src/summer/utils/ParamUtil.java
@@ -2,24 +2,31 @@ package src.summer.utils;
 
 import jakarta.servlet.http.HttpServletRequest;
 import src.summer.annotations.Param;
+import src.summer.beans.SummerSession;
+import src.summer.exception.SummerProcessException;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Enumeration;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 public abstract class ParamUtil {
     public static List<Object> getMethodParameterValues( Method method, HttpServletRequest request )
             throws NoSuchFieldException, InvocationTargetException, InstantiationException, IllegalAccessException,
-            NoSuchMethodException, ClassNotFoundException {
+            NoSuchMethodException, ClassNotFoundException, SummerProcessException {
         List<Object> values = new ArrayList<>();
         for ( Parameter parameter : method.getParameters() ) {
-            values.add( getParameterValue( parameter, request ) );
+            if ( parameter.isAnnotationPresent( Param.class ) ) {
+                values.add( getParameterValue( parameter, request ) );
+            } else {
+                if ( parameter.getType().getName().equals( SummerSession.class.getName() ) ) {
+                    HashMap<String, Object> map = SessionUtil.sessionToMap( request.getSession() );
+                    values.add( new SummerSession( map ) );
+                } else {
+                    throw new SummerProcessException( "ETU2443 - Parameters must be annotated with \"@Param\"." );
+                }
+            }
         }
         return values;
     }
@@ -33,13 +40,13 @@ public abstract class ParamUtil {
         Class<?> paramType = parameter.getType();
 
         if ( paramValue != null ) {
-            return cast( paramValue, paramType );
+            return TypeUtil.cast( paramValue, paramType );
         }
         paramValue = paramType.newInstance();
         for ( String fieldName : getFieldsNames( paramName, request ) ) {
             Field field = paramType.getDeclaredField( fieldName );
             Class<?> fieldType = field.getType();
-            Object fieldValue = cast( request.getParameter( paramName + "." + fieldName ), fieldType );
+            Object fieldValue = TypeUtil.cast( request.getParameter( paramName + "." + fieldName ), fieldType );
 
             String setterName = "set" + fieldName.substring( 0, 1 ).toUpperCase() + fieldName.substring( 1 );
             Method setterMethod = paramType.getDeclaredMethod( setterName, fieldType );
@@ -59,20 +66,5 @@ public abstract class ParamUtil {
             }
         }
         return fields;
-    }
-
-    public static Object cast( Object value, Class<?> clazz ) {
-        if ( clazz.isInstance( value ) ) {
-            return clazz.cast( value );
-        }
-
-        Optional<String> stringValue = Optional.ofNullable( ( String ) value ).map( String::valueOf );
-        return switch ( clazz.getSimpleName().toLowerCase() ) {
-            case "string" -> stringValue.orElse( null );
-            case "localdate" -> LocalDate.parse( stringValue.orElse( "" ) );
-            case "integer", "int" -> Integer.valueOf( stringValue.orElse( "0" ) );
-            case "double", "float" -> Double.valueOf( stringValue.orElse( "0.0" ) );
-            default -> throw new IllegalArgumentException( "Unsupported type: " + clazz );
-        };
     }
 }
